Moved retrieveFilePath to extension
